+++ src/Router/Router.js
@@ -12,8 +12,8 @@ import Login from '../pages/Login/Login'
 import ForgetPassword from '../pages/ForgetPassword/ForgetPassword'
 import InkController from '../pages/InkController/InkController'
 import LandingPage from '../pages/LandingPage'
-import CharacterChoicePage from '../pages/CharacterChoicePage'
-import { CharacterChoicePage2 } from '../pages/CharacterChoicePage2'
+// import CharacterChoicePage from '../pages/CharacterChoicePage'
+import { CharacterChoicePage } from '../pages/CharacterChoicePage'
 import CharacterChapterPage from '../pages/CharacterChapterPage'
 import IntroMaster from '../pages/LandingPage/IntroMaster'
 import ProfileBuilderPage from '../pages/ProfileBuilderPage'
@@ -70,7 +70,7 @@ class Router extends Component {
             </Route>
 
             <PrivateRoute path="/characterchoice" exact>
-              <CharacterChoicePage2 />
+              <CharacterChoicePage />
             </PrivateRoute>
 
             <PrivateRoute path="/chapters/:name" exact>
