+++ src/models/storyMap.js
@@ -33,6 +33,7 @@ export const CHARACTER_MAP = [
     characterId: 1,
     name: "Nadia Ibrahim",
     linkName: "nadia",
+    characterIntroImage: "/character_choice_page/nadia.png", 
     profileImage: "/images/nadia_profile.png",
     description: "Nadia wants to become a doctor - but will love and religion get in the way?",
     jsonLink: '../stories/nadia.ink.json',
@@ -123,6 +124,7 @@ export const CHARACTER_MAP = [
     characterId: 2,
     name: "Aman Singh",
     linkName: "aman",
+    characterIntroImage: "/character_choice_page/aman.png", 
     profileImage: "https://images.generated.photos/7PFNUmz7LLOsqCFQwANR09dLK1EWwS_r0XyJeh59CIQ/rs:fit:512:512/wm:0.95:sowe:18:18:0.33/Z3M6Ly9nZW5lcmF0/ZWQtcGhvdG9zL3Yz/XzAzOTI3MzIuanBn.jpg",
     description: "Aman wants to model and study overseas - but will his culture make things harder?",
     jsonLink: '../stories/aman.ink.json',
