+++ src/pages/CharacterChoicePage/constants/index.js
